+++ src/dxp/BackRequest.cpp
@@ -20,12 +20,12 @@ std::unique_ptr<MessageInterface> BackRequest::create(const std::string& message
 
 BackRequest::BackRequest(const std::string& message)
 :
-        move_number_(boost::lexical_cast<size_t>(message.substr(0, 3).c_str())),
+        move_number_(boost::lexical_cast<int>(message.substr(0, 3).c_str())),
         side_to_move_(*(message.substr(3, 1)).begin())
 {
 }
 
-size_t BackRequest::move_number() const
+int BackRequest::move_number() const
 {
         return move_number_;
 }
@@ -35,7 +35,7 @@ char BackRequest::side_to_move() const
         return side_to_move_;
 }
 
-std::string BackRequest::str(size_t m, char c)
+std::string BackRequest::str(int m, char c)
 {
         return HEADER + body(m, c);
 }
@@ -50,7 +50,7 @@ std::string BackRequest::body() const
         return body(move_number(), side_to_move());
 }
 
-std::string BackRequest::body(size_t m, char c)
+std::string BackRequest::body(int m, char c)
 {
         std::stringstream sstr;
         sstr << std::setw( 3) << std::setfill('0') << m;
