+++ src/generate/KingMoves.h
@@ -19,9 +19,9 @@ public:
         static void generate_reverse(const Position<Board>&, Stack&);
         static void generate_promotions(const Position<Board>&, Stack&);
 
-        static size_t count(const Position<Board>&);
-        static size_t count_reverse(const Position<Board>&);
-        static size_t count_promotions(const Position<Board>&);
+        static int count(const Position<Board>&);
+        static int count_reverse(const Position<Board>&);
+        static int count_promotions(const Position<Board>&);
         
         static bool detect(const Position<Board>&);
         static bool detect_reverse(const Position<Board>&);
@@ -38,19 +38,19 @@ private:
         static void generate_dirs(BitBoard, BitBoard, Stack&);
         
         // tag dispatching based on king range
-        template<size_t> static void generate_dir(BitBoard, BitBoard, Stack&);
-        template<size_t> static void generate_dir(BitBoard, BitBoard, Stack&, Int2Type<rules::RANGE_1>);
-        template<size_t> static void generate_dir(BitBoard, BitBoard, Stack&, Int2Type<rules::RANGE_N>);
+        template<int> static void generate_dir(BitBoard, BitBoard, Stack&);
+        template<int> static void generate_dir(BitBoard, BitBoard, Stack&, Int2Type<rules::RANGE_1>);
+        template<int> static void generate_dir(BitBoard, BitBoard, Stack&, Int2Type<rules::RANGE_N>);
 
-        static size_t count_dirs(BitBoard, BitBoard);
+        static int count_dirs(BitBoard, BitBoard);
         
         // tag dispatching based on king range
-        template<size_t> static size_t count_dir(BitBoard, BitBoard);
-        template<size_t> static size_t count_dir(BitBoard, BitBoard, Int2Type<rules::RANGE_1>);
-        template<size_t> static size_t count_dir(BitBoard, BitBoard, Int2Type<rules::RANGE_N>);
+        template<int> static int count_dir(BitBoard, BitBoard);
+        template<int> static int count_dir(BitBoard, BitBoard, Int2Type<rules::RANGE_1>);
+        template<int> static int count_dir(BitBoard, BitBoard, Int2Type<rules::RANGE_N>);
 
         static bool detect_dirs(BitBoard, BitBoard);
-        template<size_t> static bool detect_dir(BitBoard, BitBoard);
+        template<int> static bool detect_dir(BitBoard, BitBoard);
 };
 
 }       // namespace generate
