+++ src/generate/PawnMoves.h
@@ -17,9 +17,9 @@ public:
         static void generate_reverse(const Position<Board>&, Stack&);
         static void generate_promotions(const Position<Board>&, Stack&);
 
-        static size_t count(const Position<Board>&);
-        static size_t count_reverse(const Position<Board>&);
-        static size_t count_promotions(const Position<Board>&);
+        static int count(const Position<Board>&);
+        static int count_reverse(const Position<Board>&);
+        static int count_promotions(const Position<Board>&);
         
         static bool detect(const Position<Board>&);
         static bool detect_reverse(const Position<Board>&);
@@ -29,15 +29,15 @@ private:
         // implementation
         static void generate_dirs(BitBoard, BitBoard, Stack&);
         static void generate_reverse_dirs(BitBoard, BitBoard, Stack&);
-        template<size_t> static void generate_dir(BitBoard, BitBoard, Stack&);
+        template<int> static void generate_dir(BitBoard, BitBoard, Stack&);
 
-        static size_t count_dirs(BitBoard, BitBoard);
-        static size_t count_reverse_dirs(BitBoard, BitBoard);
-        template<size_t> static size_t count_dir(BitBoard, BitBoard);
+        static int count_dirs(BitBoard, BitBoard);
+        static int count_reverse_dirs(BitBoard, BitBoard);
+        template<int> static int count_dir(BitBoard, BitBoard);
 
         static bool detect_dirs(BitBoard, BitBoard);
         static bool detect_reverse_dirs(BitBoard, BitBoard);
-        template<size_t> static bool detect_dir(BitBoard, BitBoard);
+        template<int> static bool detect_dir(BitBoard, BitBoard);
 
         static BitBoard promotion(BitBoard);
         static BitBoard promotors(BitBoard);
